refactor: Move ObjectEntry and ObjectPage into raw
