+++ src/layers/concurrent_limit.rs
@@ -25,7 +25,6 @@ use futures::AsyncRead;
 use tokio::sync::OwnedSemaphorePermit;
 use tokio::sync::Semaphore;
 
-use crate::object::*;
 use crate::raw::*;
 use crate::*;
