+++ src/layers/logging.rs
@@ -27,7 +27,6 @@ use log::error;
 use log::trace;
 use log::warn;
 
-use crate::object::*;
 use crate::raw::*;
 use crate::*;
