+++ src/layers/subdir.rs
@@ -17,10 +17,6 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 
-use crate::object::BlockingObjectPage;
-use crate::object::BlockingObjectPager;
-use crate::object::ObjectPage;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
