+++ src/layers/tracing.rs
@@ -24,10 +24,6 @@ use async_trait::async_trait;
 use futures::AsyncRead;
 use tracing::Span;
 
-use crate::object::BlockingObjectPage;
-use crate::object::BlockingObjectPager;
-use crate::object::ObjectPage;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
