+++ src/raw/mod.rs
@@ -51,6 +51,17 @@ pub use path::validate_path;
 mod wrappers;
 pub use wrappers::apply_wrapper;
 
+mod object_entry;
+pub use object_entry::ObjectEntry;
+
+mod object_page;
+pub use object_page::BlockingObjectPage;
+pub use object_page::BlockingObjectPager;
+pub use object_page::EmptyBlockingObjectPager;
+pub use object_page::EmptyObjectPager;
+pub use object_page::ObjectPage;
+pub use object_page::ObjectPager;
+
 mod operation;
 pub use operation::Operation;
