+++ src/raw/wrappers/error_context.rs
@@ -17,9 +17,6 @@ use std::fmt::Formatter;
 
 use async_trait::async_trait;
 
-use crate::object::BlockingObjectPager;
-use crate::object::ObjectPage;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
