+++ src/services/azblob/backend.rs
@@ -31,7 +31,6 @@ use reqsign::AzureStorageSigner;
 use super::dir_stream::DirStream;
 use super::error::parse_error;
 use crate::object::ObjectMetadata;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
