+++ src/services/azblob/dir_stream.rs
@@ -22,7 +22,6 @@ use time::OffsetDateTime;
 
 use super::backend::Backend;
 use super::error::parse_error;
-use crate::object::ObjectPage;
 use crate::raw::*;
 use crate::*;
