+++ src/services/ftp/backend.rs
@@ -38,7 +38,6 @@ use tokio::sync::OnceCell;
 use super::dir_stream::DirStream;
 use super::dir_stream::ReadDir;
 use super::util::FtpReader;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
