+++ src/services/gcs/dir_stream.rs
@@ -23,7 +23,6 @@ use time::OffsetDateTime;
 use super::backend::Backend;
 use super::error::parse_error;
 use super::error::parse_json_deserialize_error;
-use crate::object::ObjectPage;
 use crate::raw::*;
 use crate::*;
