+++ src/services/hdfs/backend.rs
@@ -27,8 +27,6 @@ use time::OffsetDateTime;
 
 use super::dir_stream::DirStream;
 use super::error::parse_io_error;
-use crate::object::EmptyObjectPager;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
