+++ src/services/ipfs/backend.rs
@@ -25,8 +25,6 @@ use prost::Message;
 
 use super::error::parse_error;
 use super::ipld::PBNode;
-use crate::object::ObjectPage;
-use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
