+++ src/services/ipmfs/dir_stream.rs
@@ -21,9 +21,7 @@ use serde::Deserialize;
 use super::backend::Backend;
 use super::error::parse_error;
 use super::error::parse_json_deserialize_error;
-use crate::object::ObjectPage;
 use crate::raw::*;
-use crate::ObjectEntry;
 use crate::ObjectMetadata;
 use crate::ObjectMode;
 use crate::Result;
