+++ src/services/obs/backend.rs
@@ -26,10 +26,9 @@ use http::Uri;
 use log::debug;
 use reqsign::HuaweicloudObsSigner;
 
+use super::dir_stream::DirStream;
 use super::error::parse_error;
-use crate::object::ObjectPager;
 use crate::raw::*;
-use crate::services::obs::dir_stream::DirStream;
 use crate::*;
 
 /// Builder for Huaweicloud OBS services
