+++ src/services/s3/dir_stream.rs
@@ -24,11 +24,9 @@ use time::OffsetDateTime;
 use super::backend::Backend;
 use super::error::parse_error;
 use super::error::parse_xml_deserialize_error;
-use crate::object::ObjectPage;
 use crate::raw::*;
 use crate::Error;
 use crate::ErrorKind;
-use crate::ObjectEntry;
 use crate::ObjectMetadata;
 use crate::ObjectMode;
 use crate::Result;
